Remove write_data_transformer
This is being changed to prepare for the MOVK and MOVZ instructions. The
ALU result signal would have needed to be routed to the
data_transformer, effectively nullifying the need for the mux3. It is
more preferrable to just get rid of the mux and do everything in the
top-level Writeback module, since the logic is fairly simple.