Refacted ALU_op/ALU_control
The ALU control module was a remnant from the ELC3338 course, and one
which I always found useless. It took a signal from the control module,
as well as the opcode, in order to make its decision on which mode the
ALU should operate in.

If it takes the opcode, as well as a signal from the control module,
which also accepts the opcode, than it doesn't have any more information
than the control module does. Because of this, the module is pointless,
and the alu mode might as well be determined in the control module.

The control module now outputs a 4-bit signal named alu_op, which the
execute module accepts and passes to the ALU for determining its mode of
operation.