Actually finished multiplier
I crunched out some 8 bit multiplications on paper and actually learned
how interpreting the multiplicand & multiplier as unsigned or signed
actually affects the result. I then implemented this knowledge into the
multiplier. Using the previous Booth 4-radix multiplier found online, it
was a simple fix: Sign extend the multiplicand with a 0 if unsigned, or
the MSB if signed. The algorithm takes care of the rest.